Create Main.py
The blue square is the "player" and the red square is the "goal". The player and the goal are both 20 pixels by 20 pixels. 

The player should start in the top left corner of the world. You can place the goal anywhere if its x and y values are both multiples of 20. Here is a reasonable configuration where the goal is at (360, 360):